Ability to modify ticketing options
I have a case whereby we would offer postal and box office tickets. Because postal tickets would take about 4 or 5 days to get to an attendee, it would be preferable just to disable the postal option but keep the box office option going until the expiry date set in the event. With that in mind, I have added a filter "mt_ticketing_availability" which filters the available ticketing options with a view to modifying it. The filter also accepts the $cart variable so that checks can be made against the cart items and the ticketing amended appropriately.