Remove logic that limited cart updates based on expiration time
This contained a logic fault if the user's expiration time was *later* than the current time plus the current expiration setting, which would happen if the cart expiration time was changed while cart's were active. Logic was actually irrelevant, since there is no need to limit expiration updates and cart updates based on when carts expire.